% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_to_RNA.R
\name{DNA_to_RNA}
\alias{DNA_to_RNA}
\title{DNA_to_RNA}
\usage{
DNA_to_RNA(DNA_sequence)
}
\arguments{
\item{DNA_sequence}{Sequence of the nucleotides A, C, G and T}
}
\value{
Sequence of the nucleotides A, C, G and U corresponding to the DNA sequence given as parameter
}
\description{
DNA_to_RNA
}
\examples{
DNA_to_RNA("A C G T")
}
