% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_Codon.R
\name{Generate_Codon}
\alias{Generate_Codon}
\title{Generate_Codon}
\usage{
Generate_Codon(enter_gene, start = 1)
}
\arguments{
\item{enter_gene}{enter the genome to create codon}

\item{start}{the start character for codon creation}
}
\value{
the function will generate partitions of codons based on the gene that has inserted
}
\description{
Generate_Codon
}
\examples{
Generate_Codon("GTCGTC")
}
