% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_rand_DNA.R
\name{Generate_rand_DNA}
\alias{Generate_rand_DNA}
\title{Generate_rand_DNA}
\usage{
Generate_rand_DNA(DNA_len)
}
\arguments{
\item{DNA_len}{insert the length of the random generated DNA}
}
\value{
The function will return a random DNA code with by the parameter length
}
\description{
Generate_rand_DNA
}
\examples{
Generate_rand_DNA(3)
}
